done split index
